requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
mysql-connector-python>=8.1.0
PyMySQL
Flask-Cors==4.0.0
Werkzeug==3.0.1
requests==2.31.0
pytest==8.2.1
pytest-cov==5.0.0